selfmastery/requirements.txt
# Web框架和API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 桌面应用框架
PyQt6==6.6.1
PyQt6-tools==6.6.0

# 数据库和ORM
SQLAlchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
aiosqlite==0.19.0

# 任务队列和缓存
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# HTTP客户端
httpx==0.25.2
requests==2.31.0

# 认证和安全
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# 配置管理
python-dotenv==1.0.0
pyyaml==6.0.1

# 日志和监控
loguru==0.7.2
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# 测试框架
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2

# 开发工具
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# 数据处理
pandas==2.1.4
numpy==1.25.2
openpyxl==3.1.2

# 时间处理
python-dateutil==2.8.2
pytz==2023.3

# 其他工具
click==8.1.7
rich==13.7.0
typer==0.9.0